# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API REST Hoteles",
        "description": "API REST Hoteles",
        "contact": {
            "email": "cesarpozo25@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost/apirest-hoteles/public/",
    "basePath": "",
    "paths": {
...
        "/hotets": {
            "get": {
                "summary": "Get Hoteles",
                "parameters": [
                    {
                        "name": "Autorization",
                        "in": "header",
                        "description": "Bearer {token_access}",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#definitions/Hotel"
                            }
                        }
                    },
                    "403": {
                        "description": "Autorization Exception",
                        "schema": {
                            "$ref": "#definitions/Errors403"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#definitions/Errors500"
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/hotels": {
            "post": {
                "summary": "Create Hotel for store",
                "parameters": [
                    {
                        "name": "Autorization",
                        "in": "header",
                        "description": "Bearer {token_access}",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "nombre": {
                                    "type": "string",
                                    "example": "nombre apellido"
                                },
                                "NIF": {
                                    "type": "string",
                                    "example": "12345678Z"
                                },
                                "Localidad_id": {
                                    "type": "integer",
                                    "example": 1
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Create successful operation",
                        "schema": {
                            "$ref": "#definitions/Hotel"
                        }
                    },
                    "403": {
                        "description": "Autorization Exception",
                        "schema": {
                            "$ref": "#definitions/Errors403"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#definitions/Errors500"
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/hotels/{hotel_id}": {
            "get": {
                "summary": "Get Hotel",
                "parameters": [
                    {
                        "name": "hotel_id",
                        "in": "path",
                        "description": "un numero id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Autorization",
                        "in": "header",
                        "description": "Bearer {token_access}",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#definitions/Hotel"
                            }
                        }
                    },
                    "403": {
                        "description": "Autorization Exception",
                        "schema": {
                            "$ref": "#definitions/Errors403"
                        }
                    },
                    "404": {
                        "description": "Not Found Exception",
                        "schema": {
                            "$ref": "#definitions/Errors404"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#definitions/Errors500"
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "put": {
                "summary": "Update hotel",
                "parameters": [
                    {
                        "name": "Autorization",
                        "in": "header",
                        "description": "Bearer {token_access}",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "hotel_id",
                        "in": "path",
                        "description": "un numero id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "data",
                        "in": "body",
                        "required": false,
                        "schema": {
                            "properties": {
                                "nombre": {
                                    "type": "string",
                                    "example": "nombre apellido"
                                },
                                "NIF": {
                                    "type": "string",
                                    "example": "12345678Z"
                                },
                                "Localidad_id": {
                                    "type": "integer",
                                    "example": 1
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Update successful operation",
                        "schema": {
                            "$ref": "#definitions/Hotel"
                        }
                    },
                    "403": {
                        "description": "Autorization Exception",
                        "schema": {
                            "$ref": "#definitions/Errors403"
                        }
                    },
                    "404": {
                        "description": "Not Found Exception",
                        "schema": {
                            "$ref": "#definitions/Errors404"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#definitions/Errors500"
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete hotel",
                "parameters": [
                    {
                        "name": "hotel_id",
                        "in": "path",
                        "description": "un numero id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Autorization",
                        "in": "header",
                        "description": "Bearer {token_access}",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#definitions/Hotel"
                            }
                        }
                    },
                    "403": {
                        "description": "Autorization Exception",
                        "schema": {
                            "$ref": "#definitions/Errors403"
                        }
                    },
                    "404": {
                        "description": "Not Found Exception",
                        "schema": {
                            "$ref": "#definitions/Errors404"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#definitions/Errors500"
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
...
}